US009 - Refactor DS | Include DTO, Mapper, factory and service US009 - Implement UserStoryService US009 - Implement OutputUsDTO US009: Refactoring: CreateUserStoryController, class UserStory, UserStoryStore, UserStoryDTO (resolve #137)  (resolve #139)  (resolve #140)  (resolve #141)

+++ src/main/java/switch2021/project/dto/UserStoryStatusDTO.java
@@ -24,7 +24,7 @@ public class UserStoryStatusDTO {
          * Constructor to test (without SINGLETON)
          **/
 
-        public UserStoryStatusDTO(String userStoryId,String title, String description, int timeEstimate, int priority) {
+        public UserStoryStatusDTO(String userStoryId,String title, String description, double timeEstimate, int priority) {
                 this.title = new UsTitle(title);
                 this.description = new Description(description);
                 this.timeEstimate = new UsHour(timeEstimate);
